+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added top-level `pgvector` package
 - Added support for pg8000
+- Added support for `bytes` to `Bit` constructor
 - Changed `globally` option to default to `False` for Psycopg 2
 - Changed `arrays` option to default to `True` for Psycopg 2
 - Fixed equality for `Vector`, `HalfVector`, `Bit`, and `SparseVector` classes